Change to PluginInterface - prerequisites now correctly use Int64 as the version number instead of Integer which it was previously.

+++ FruitTartInterface/Network/FruitTart/PluginInterface.hs
@@ -53,7 +53,7 @@ data Interface = Interface {
       interfaceModuleName :: String,
       interfaceModuleVersion :: Int64,
       interfaceModuleSchemaVersion :: Int64,
-      interfacePrerequisites :: [(String, Integer)],
+      interfacePrerequisites :: [(String, Int64)],
       interfaceInitDatabase :: Database -> IO (),
       interfaceImportFunctionTableMVar :: MVar CombinedFunctionTable
     }
